+++ docs/versionhistory.rst
@@ -36,6 +36,8 @@ This library adheres to `Semantic Versioning 2.0 <http://semver.org/>`_.
 
 **4.14.0**
 
+- Added an implementation of the ``asyncio.Future`` object for single pending objects.
+  (`#1146 <https://github.com/agronholm/anyio/pull/1146>`_; PR by @Vizonex)
 - Added support for Python 3.15
 - Added an asynchronous implementation of the ``itertools`` module
   (`#998 <https://github.com/agronholm/anyio/issues/998>`_; PR by @11kkw)